Removed ARM 64-bit support
Removed ARM 64-bit support from build.common.gradle because
"We believe the problem occurred when native support for 64-bit devices was introduced with the v5.2 release.
In August 2019, the Google Play store changed its requirements so that it will only allow apps in the store that include native support for 64-bit devices.
We believe that there is an external library that is incorporated into the UVC webcam portion of our software that is not 64-bit compliant.
We suspect that this results in a memory corruption error when an external webcam is used with the v5.2 software.
This bug did not manifest itself in the pre-release testing and is currently only present in the release version of the v5.2 software."
- FTC Wiki
https://github.com/FIRST-Tech-Challenge/skystone/wiki/Troubleshooting#problem-with-uvc-webcam-support-and-version-52-software-listed-092019